packages/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api"
dynamic = ["version"]
description = "FastAPI backend application"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Generated by MortgageAI CLI" },
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.2.3",
    "sqladmin>=0.16.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.25.0",
    "minio>=7.2.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "psycopg2-binary>=2.9.0",

    "mortgage-ai-db",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/example/mortgage-ai"
Repository = "https://github.com/example/mortgage-ai.git"

[tool.hatch.version]
path = "src/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = [".", "tests"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
"mortgage-ai-db" = { path = "../db", editable = true }
